requirements.txt
catboost==1.2.2
matplotlib==3.7.3
numpy==1.24.4
pandas==1.2.5
scikit-learn==1.3.2
scipy==1.10.1
python==3.8.8
pytorch==1.8.1
xgboost==2.0.1
